combine deterineCredits into calculateCredits util file
